+++ src/main/java/com/grad/project/nc/model/ProductCharacteristicValue.java
@@ -4,31 +4,31 @@ package com.grad.project.nc.model;
  * Created by Alex on 4/24/2017.
  */
 public class ProductCharacteristicValue {
-    private int productId;
-    private int productCharacteristicId;
-    private int valueId;
+    private Long productId;
+    private Long productCharacteristicId;
+    private Long valueId;
 
-    public int getProductId() {
+    public Long getProductId() {
         return productId;
     }
 
-    public void setProductId(int productId) {
+    public void setProductId(Long productId) {
         this.productId = productId;
     }
 
-    public int getProductCharacteristicId() {
+    public Long getProductCharacteristicId() {
         return productCharacteristicId;
     }
 
-    public void setProductCharacteristicId(int productCharacteristicId) {
+    public void setProductCharacteristicId(Long productCharacteristicId) {
         this.productCharacteristicId = productCharacteristicId;
     }
 
-    public int getValueId() {
+    public Long getValueId() {
         return valueId;
     }
 
-    public void setValueId(int valueId) {
+    public void setValueId(Long valueId) {
         this.valueId = valueId;
     }
 }
