+++ src/main/java/com/grad/project/nc/persistence/mappers/ProductOrderRowMapper.java
@@ -11,12 +11,12 @@ public class ProductOrderRowMapper implements RowMapper<ProductOrder> {
     public ProductOrder mapRow(ResultSet resultSet, int i) throws SQLException {
         ProductOrder productOrder = new ProductOrder();
 
-        productOrder.setProductOrderId(resultSet.getInt("product_order_id"));
-        productOrder.setProductInstanceId(resultSet.getInt("product_instance_id"));
-        productOrder.setUserId(resultSet.getInt("user_id"));
-        productOrder.setCategoryId(resultSet.getInt("category_id"));
-        productOrder.setStatusID(resultSet.getInt("status_id"));
-        productOrder.setResponsibleId(resultSet.getInt("responsible_id"));
+        productOrder.setProductOrderId(resultSet.getLong("product_order_id"));
+        productOrder.setProductInstanceId(resultSet.getLong("product_instance_id"));
+        productOrder.setUserId(resultSet.getLong("user_id"));
+        productOrder.setCategoryId(resultSet.getLong("category_id"));
+        productOrder.setStatusID(resultSet.getLong("status_id"));
+        productOrder.setResponsibleId(resultSet.getLong("responsible_id"));
         productOrder.setOpenDate(resultSet.getTimestamp("open_date").toLocalDateTime());
         productOrder.setCloseDate(resultSet.getTimestamp("close_date").toLocalDateTime());
